Parameters: New parameter system.
Now, instead of just vector of doubles and constrainment vector, we have simple parameter system.
It allows to have different parameters type, has atomic values (should be threadsafe),
and has built-in serialization into JSON objects.

Also, it's should be possible to make a simple Undo/Redo events manager, per-voice parameter distribution,
and other cool things.

+++ dafx/tubedist.hpp
@@ -42,8 +42,8 @@ public:
 	void process_params() override;
 	void process_callback() override;
 private:
-	double tC = 0, bC = 0, pC = 0;
-	double pre_gain = 0, gain = 0, out_gain = 0;
+	float tC = 0, bC = 0, pC = 0;
+	float pre_gain = 0, gain = 0, out_gain = 0;
 };
 }
 }
